Preserve node arguments with spaces
Previously, arguments with spaces in them would get re-split before being
passed to node.  So, for instance, the following command:

    magic-ws -p "/path/to/project/*" node /path/to/bin.js --arg "with a space"

would result in running bin.js with the following contents in argv:

    ['/path/to/node', '/path/to/bin.js', '--arg', 'with', 'a', 'space']

With this change, bin.js will receive the correct argv:

    ['/path/to/node', '/path/to/bin.js', '--arg', 'with a space']

Note that this change requires the script to use /bin/bash instead of /bin/sh.